tweak long to decimal

+++ modules/No.StoreButler.LedgerManagement/src/No.StoreButler.LedgerManagement.Application.Contracts/LedgerManagement/Dtos/CreatePayTradeInput.cs
@@ -13,7 +13,7 @@ namespace No.StoreButler.LedgerManagement.LedgerManagement.Dtos
         /// <summary>
         /// 金额
         /// </summary>
-        public long PayAmount { get; set; }
+        public decimal PayAmount { get; set; }
 
         /// <summary>
         /// 资金流向
